requirements.txt
asgiref==3.6.0
dj-database-url==1.2.0
Django==4.1.7
django-filter==22.1
django-sendsms==0.5
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
drfaddons==0.1.4
gunicorn==20.1.0
Pillow==9.4.0
psycopg2==2.9.5
PyJWT==1.7.1
pytz==2022.7.1
sqlparse==0.4.3
whitenoise==6.3.0
MarkupSafe==2.1.2
certifi==2022.12.7
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
jinja2==3.1.2
packaging==23.0
requests==2.28.2
ruamel.yaml==0.17.21
uritemplate==4.1.1
urllib3==1.26.14
django_heroku==0.3.1
python-http-client==3.3.7
sendgrid==6.9.7
starkbank-ecdsa==2.2.0
configparser==5.3.0


django-dual-authentication==1.2.1
# oauth2_provider==0.0
# drf_social_oauth2==1.2.1

# asgi
Automat==22.10.0
attrs==22.2.0
autobahn==23.1.2
cffi==1.15.1
constantly==15.1.0
cryptography==39.0.1
daphne==4.0.0
hyperlink==21.0.0
incremental==22.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyopenssl==23.0.0
service-identity==21.1.0
six==1.16.0
twisted==22.10.0
txaio==23.1.1
typing-extensions==4.5.0
zope.interface==5.5.2

# websocket/channels
channels==4.0.0
websockets==10.4
websocket-client==1.5.1

django-phonenumber-field==7.0.2
phonenumberslite==8.13.6
phonenumbers==8.13.6
Babel==2.11.0

